backend/requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
scikit-learn
nltk
lxml
lxml_html_clean
newspaper3k